refactor: 어드민 폴더 구조 정리
