+++ .gitignore
@@ -27,3 +27,5 @@ dist-ssr
 
 *storybook.log
 storybook-static
+
+.claude
